requirements.txt
tkinter
Pillow
requests
pyinstaller